+++ build.gradle.kts
@@ -1,0 +1,52 @@
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+plugins {
+    kotlin("jvm") version "1.6.0-M1"
+}
+
+dependencies {
+    api("junit:junit:4.13")
+    implementation("junit:junit:4.13")
+    testImplementation("junit:junit:4.13")
+    implementation(kotlin("stdlib-jdk8"))
+}
+
+configurations {
+    implementation {
+        resolutionStrategy.failOnVersionConflict()
+    }
+}
+
+sourceSets {
+    main {
+        java.srcDir("src/main/kotlin")
+    }
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+
+tasks {
+    test {
+        testLogging.showExceptions = true
+    }
+}
+
+dependencies {
+
+}
+repositories {
+    mavenCentral()
+}
+
+
+val compileKotlin: KotlinCompile by tasks
+compileKotlin.kotlinOptions {
+    jvmTarget = JavaVersion.VERSION_11.toString()
+}
+val compileTestKotlin: KotlinCompile by tasks
+compileTestKotlin.kotlinOptions {
+    jvmTarget = JavaVersion.VERSION_11.toString()
+}
