+++ src/main/kotlin/Factory.kt
@@ -1,7 +1,5 @@
 package main.kotlin
 
-import org.junit.Test
-
 sealed class Country
 
 object Spain : Country()
@@ -21,19 +19,4 @@ object CurrencyFactory {
             is Canada -> Currency("CAD")
             is Poland -> Currency("PLN")
         }
-}
-
-class FactoryMethodTest {
-    @Test
-    fun currencyTest() {
-        val usaCurrency = CurrencyFactory.currencyForCountry(USA).code
-        println("USA currency: $usaCurrency")
-
-        val greekCurrency = CurrencyFactory.currencyForCountry(Greece).code
-        println("Greek currency: $greekCurrency")
-
-        assert(greekCurrency == "EUR")
-        assert(usaCurrency == "USD")
-
-    }
 }
