--- src/main/kotlin/Root.kt
@@ -1,4 +1,0 @@
-package main.kotlin
-
-class Root {
-}
