+++ src/main/kotlin/Singleton.kt
@@ -1,22 +1,9 @@
 package main.kotlin
 
-import org.junit.Test
-
 object NetworkDriver {
     init {
         println("Initializing: $this")
     }
 
     fun log(): NetworkDriver = apply { println("Network driver: $this") }
-}
-
-class SingletonTest {
-    @Test
-    fun testSingleton() {
-        println("Start")
-        val networkDriver1 = NetworkDriver.log()
-        val networkDriver2 = NetworkDriver.log()
-
-        assert(networkDriver1 === networkDriver2)
-    }
 }
